fix(complexities): change float32 to float64 for all result tensors
This avoids a potential error when writing results to a parquet file.
Specifying the complexity metrics in an order in which a float16 tensor
is appended to a float32 tensor in a dataframe and then written to
parquet results in an error as downcasting types seems not possible here.
To be shure to have optimal precision float64 is chosen for all metrics.